+++ packages/vibrant-core/src/lib/nl2br/index.ts
@@ -1,0 +1,1 @@
+export { nl2br } from './nl2br';