+++ packages/workspace-plugin/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.31.1](https://github.com/pedaling/opensource/compare/workspace-plugin-0.31.0...workspace-plugin-0.31.1) (2023-02-28)
+
+## [0.31.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.30.0...workspace-plugin-0.31.0) (2023-02-28)
+
+## [0.30.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.29.1...workspace-plugin-0.30.0) (2023-02-24)
+
+### [0.29.1](https://github.com/pedaling/opensource/compare/workspace-plugin-0.29.0...workspace-plugin-0.29.1) (2023-02-23)
+
 ## [0.29.0](https://github.com/pedaling/opensource/compare/workspace-plugin-0.28.12...workspace-plugin-0.29.0) (2023-02-23)
 
 ### [0.28.12](https://github.com/pedaling/opensource/compare/workspace-plugin-0.28.11...workspace-plugin-0.28.12) (2023-02-21)